check: latency timer + reformatting

+++ src/api/FTDI_Device_setBaudRate.c
@@ -23,7 +23,7 @@ napi_value device_setBaudRate(napi_env env, napi_callback_info info) {
   if(utils_check(napi_get_value_uint32(env, argv[0], &baud_rate), "Baud rate must be a number", "wrongarg")) return NULL;
 
   // Update FTDI device
-  utils_check(FT_SetBaudRate(instance_data->ftHandle, baud_rate));
+  utils_check(FT_SetBaudRate(instance_data->ftHandle, (ULONG)baud_rate));
 
   return NULL;
 }
